Fix/lock date and voting power (#233)
