+++ src/lib/helpers/locking.ts
@@ -5,6 +5,8 @@ import {
   eachDayOfInterval,
   isWednesday,
   nextWednesday,
+  startOfWeek,
+  subWeeks,
 } from "date-fns";
 
 export default abstract class LockingHelper {
@@ -52,4 +54,28 @@ export default abstract class LockingHelper {
       ...this.getDaysExceptWednesday(),
     ];
   };
+
+  public static calculateExpirationDate(
+    currentWeek: number,
+    weekLocked: number,
+    cliff: number,
+    slope: number,
+  ): Date {
+    // Calculate weeks passed since lock
+    const weeksPassed = currentWeek - weekLocked;
+
+    // Calculate remaining weeks in the vesting schedule
+    const remainingWeeks = cliff + slope - weeksPassed;
+
+    // Calculate the initial lock date by subtracting weeks passed from the start of the current week
+    const startOfCurrentWeek = startOfWeek(new Date(), { weekStartsOn: 3 });
+    const initialLockDate = nextWednesday(
+      subWeeks(startOfCurrentWeek, weeksPassed),
+    );
+
+    // Add the remaining weeks to the initial lock date
+    let expirationDate = addWeeks(initialLockDate, remainingWeeks);
+
+    return expirationDate;
+  }
 }
